rootfs/requirements.txt
google-cloud-secret-manager
grpcio==1.57.0
urllib3
